refactor: Rename AuthActions to LoginActions
